refactor: implement Editor component
